requirements.txt
customtkinter
discord.py-self
requests
psutil
pyautogui
